change cover size in emc for "large cover" plus 25%

+++ usr/lib/enigma2/python/Plugins/Extensions/MyMetrixLite/MainSettingsView.py
@@ -1136,7 +1136,7 @@ class MainSettingsView(Screen):
                         move(dpath + file,dpath + file + ".hd")
                     copy(spath + file,dpath + file)
                 else:
-                    if not path.exists(dpath + file + ".del"):
+                    if not path.exists(dpath + file + ".hd") and not path.exists(dpath + file + ".del"):
                         f = open(dpath + file + ".del", "w")
                         f.close()
                     copy(spath + file,dpath + file)
